Fix reload and unify dumpdata fixtures
