--- journal/fixtures/auth_users.json
@@ -1,132 +1,0 @@
-[
-{
-  "pk": 1,
-  "fields": {
-    "groups": [
-      1
-    ],
-    "email": "kevinlee963@gmail.com",
-    "password": "!TfhVfBl5PPgzeZUrgrGCbexeVfKpqOkz9TqKio2j",
-    "date_joined": "2015-07-24T11:06:17.433",
-    "last_login": "2015-07-28T15:19:04.176",
-    "is_superuser": false,
-    "first_name": "Kevin",
-    "last_name": "Lee",
-    "username": "kevinlee963@gmail.com",
-    "user_permissions": [
-      76
-    ],
-    "is_staff": false,
-    "is_active": true
-  },
-  "model": "auth.user"
-},
-{
-  "pk": 2,
-  "fields": {
-    "groups": [
-      1
-    ],
-    "email": "janicez348@gmail.com",
-    "password": "!4GUDC8tm0slriViUOH6ZIeIjyRlb74Qg7wVfTDLk",
-    "date_joined": "2015-07-24T11:06:48.118",
-    "last_login": null,
-    "is_superuser": false,
-    "first_name": "",
-    "last_name": "",
-    "username": "janicez348@gmail.com",
-    "user_permissions": [
-      76
-    ],
-    "is_staff": false,
-    "is_active": true
-  },
-  "model": "auth.user"
-},
-{
-  "pk": 3,
-  "fields": {
-    "groups": [
-      1
-    ],
-    "email": "leekevin@berkeley.edu",
-    "password": "!GDliRjaNKcK4vnoI9mi922WykLBkKXH4UIKHWZGi",
-    "date_joined": "2015-07-24T11:07:07.653",
-    "last_login": null,
-    "is_superuser": false,
-    "first_name": "",
-    "last_name": "",
-    "username": "leekevin@berkeley.edu",
-    "user_permissions": [
-      77
-    ],
-    "is_staff": true,
-    "is_active": true
-  },
-  "model": "auth.user"
-},
-{
-  "pk": 6,
-  "fields": {
-    "groups": [
-      1
-    ],
-    "email": "janicezhao348@gmail.com",
-    "password": "!fLHVspFzqoVd6JkTl0zAtjnGRnXfG7q7Wka9vgdO",
-    "date_joined": "2015-07-24T21:44:34.121",
-    "last_login": null,
-    "is_superuser": false,
-    "first_name": "",
-    "last_name": "",
-    "username": "janicezhao348@gmail.com",
-    "user_permissions": [
-      77
-    ],
-    "is_staff": false,
-    "is_active": true
-  },
-  "model": "auth.user"
-},
-{
-  "pk": 7,
-  "fields": {
-    "groups": [],
-    "email": "comet.tester@gmail.com",
-    "password": "!1WBaSr0IhPHt9aLzyBgCOOYdM1WI5pqkPKJcIpxJ",
-    "date_joined": "2015-07-26T11:25:47.463",
-    "last_login": "2015-07-27T10:33:00.288",
-    "is_superuser": false,
-    "first_name": "Comet",
-    "last_name": "Test",
-    "username": "comet.tester@gmail.com",
-    "user_permissions": [
-      76
-    ],
-    "is_staff": false,
-    "is_active": true
-  },
-  "model": "auth.user"
-},
-{
-  "pk": 8,
-  "fields": {
-    "groups": [
-      1
-    ],
-    "email": "kevinlee9630@gmail.com",
-    "password": "!PMd3535u3yvDNUnw69ClJ3gfHmzAjIsNvbrbw670",
-    "date_joined": "2015-07-28T17:59:18.099",
-    "last_login": null,
-    "is_superuser": false,
-    "first_name": "",
-    "last_name": "",
-    "username": "kevinlee9630@gmail.com",
-    "user_permissions": [
-      78
-    ],
-    "is_staff": false,
-    "is_active": true
-  },
-  "model": "auth.user"
-}
-]
