--- journal/fixtures/groups.json
@@ -1,10 +1,0 @@
-[
-{
-  "pk": 1,
-  "fields": {
-    "name": "Glen A. Wilson HS",
-    "permissions": []
-  },
-  "model": "auth.group"
-}
-]
