+++ journal/fixtures/groups.yaml
@@ -1,0 +1,5 @@
+- fields:
+    name: Glen A. Wilson HS
+    permissions: []
+  model: auth.group
+  pk: 1
