+++ journal/tests/test_activity.py
@@ -13,6 +13,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
 from selenium.webdriver.common.by import By
 from django.contrib.staticfiles.testing import StaticLiveServerTestCase
 from selenium.webdriver.firefox.webdriver import WebDriver
@@ -26,15 +27,16 @@ from django.core.management import call_command
 
 class ActivitySeleleniumTests(StaticLiveServerTestCase):
     """Selenium tests for the activity page"""
-    fixtures = ['tests/fixtures/init_data.json']
 
     def __init__(self, *args, **kwargs):
-        super(ActivitySeleleniumTests, self).__init__(*args, **kwargs)
+        super().__init__(*args, **kwargs)
         if not settings.DEBUG:
             settings.DEBUG = True
 
     def setUp(self):
         """Handles login and things"""
+        call_command('flush', interactive=False, verbosity=0)  # Clears db
+        call_command('loaddata', 'test_init', commit=False, verbosity=0)
         self.selenium = WebDriver()
         self.selenium.set_window_size(1024, 800)
         self.selenium.get('{0}/{1}'.format(self.live_server_url, ''))
@@ -52,6 +54,44 @@ class ActivitySeleleniumTests(StaticLiveServerTestCase):
     def tearDown(self):
         self.selenium.quit()
         super()
+
+    def test_activity_form_back(self):
+        """make sure the back button works"""
+        self.selenium\
+            .find_element_by_xpath("//img[@src='/static/journal/activities/img"
+                                   "/journal_sign.png']")
+        self.selenium.find_element_by_xpath("/html/body/div[1]/div[2]/div/div[1]/div/div[2]/div/a").click()
+        self.selenium.find_element_by_name('activity_name').send_keys('Walking the cat')
+        self.selenium.find_element_by_name('activity_description').send_keys('Walking the cat around the neighborhood')
+        self.selenium.find_element_by_xpath('/html/body/div[1]/div[2]/div[1]/form/div[2]/div[2]/input').send_keys('02/07/1990')
+        self.selenium.find_element_by_xpath('//*[@id="id_activity_type"]/li[2]/label').click()
+        self.selenium.find_element_by_xpath('//*[@id="id_activity_type"]/li[3]/label').click()
+        self.selenium.find_element_by_xpath('//*[@id="id_learned_objective"]/li[6]/label').click()
+        self.selenium.find_element_by_name('activity_adviser').send_keys('Cat')
+        self.selenium.find_element_by_name('advisor_phone').send_keys('1234567890')
+        self.selenium.find_element_by_name('advisor_email').send_keys('kitty@cats.cat')
+        self.selenium.find_element_by_link_text('Back').click()
+        self.selenium\
+            .find_element_by_xpath("//img[@src='/static/journal/activities/img"
+                                   "/journal_sign.png']")
+
+    def test_activity_form_error(self):
+        """Tests to check errors on the activity form"""
+        self.selenium\
+            .find_element_by_xpath("//img[@src='/static/journal/activities/img"
+                                   "/journal_sign.png']")
+        self.selenium.find_element_by_xpath("/html/body/div[1]/div[2]/div/div[1]/div/div[2]/div/a").click()
+        # self.selenium.find_element_by_name('activity_name').send_keys('')
+        self.selenium.find_element_by_name('activity_description').send_keys('Walking with huahua around the neighborhood')
+        self.selenium.find_element_by_xpath('/html/body/div[1]/div[2]/div[1]/form/div[2]/div[2]/input').send_keys('02/07/1990')
+        self.selenium.find_element_by_xpath('//*[@id="id_activity_type"]/li[2]/label').click()
+        self.selenium.find_element_by_xpath('//*[@id="id_activity_type"]/li[3]/label').click()
+        self.selenium.find_element_by_xpath('//*[@id="id_learned_objective"]/li[6]/label').click()
+        self.selenium.find_element_by_name('activity_adviser').send_keys('Cat')
+        self.selenium.find_element_by_name('advisor_phone').send_keys('1234567890')
+        self.selenium.find_element_by_name('advisor_email').send_keys('kitty@cats.cat')
+        self.selenium.find_element_by_xpath('/html/body/div[1]/div[2]/div[1]/form/div[6]/div/input').click()
+        self.selenium.find_element_by_name('activity_description').text
 
     def test_activity_form(self):
         """Tests to ensure that activities page has all necessary elements."""
